Fix #265: Add campfires as a heat source
This pull request adds support for distinguishing if a block is a heat source not only by its resource location, but also by its state. For example, a lit campfire should provide heat to a crucible, but an unlit one should not. This was not previously possible.

This PR also adds lit furnaces and redstone torches to the list of available heat sources